+++ src/App.module.scss
@@ -12,10 +12,6 @@ flex-direction: center;
 align-items: center;
 }
 
-div {
-  height: 100%;
-  margin: 0 auto;
-}
