+++ src/components/travelInfo/RouteLogin/RouteLogin.module.scss
@@ -1,3 +1,19 @@
+.RouteLoginDiv {
+  background: #247291;
+  border-radius: 10px;
+  width: auto;
+  max-width: 100%;
+  padding-bottom: 20px;
+}
+
+.HeaderOne{
+  padding: 2%;
+  border-radius: 10px 10px 0 0;
+  background:#F7D95B; 
+  font-size: 5rem;
+  color: #247291;
+}
+
 .Link {
   color: white;
   text-decoration: none;
