+++ src/components/travelInfo/SearchForm/SearchForm.module.scss
@@ -1,0 +1,22 @@
+.SearchForm {
+
+}
+
+.SearchInput {
+    font-size: 1.3rem;
+    border-radius: 6px;
+    border: .5px solid black;
+    height: 2.5rem;
+}
+
+.SearchBtn {
+    color: #247291;
+    background: #F7D95B;
+    text-transform: uppercase;
+    padding: 5px 15px;
+    border-radius: 10px;
+    border: none;
+    text-shadow: 0 1px 1px rgba(0, 0, 0, 0.5);
+    box-shadow: 0 1px 1px rgba(0, 0, 0, 0.5);
+    margin: 1rem .7rem; 
+}
