+++ src/components/travelInfo/TravelCard/TravelCard.module.scss
